+++ facebook_auth/test/flutter_facebook_auth_test.dart
@@ -4,11 +4,10 @@ import 'package:flutter_facebook_auth_platform_interface/flutter_facebook_auth_p
 import 'package:flutter_test/flutter_test.dart';
 import 'src/data.dart';
 
-bool isLogged = false;
-
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
   group('authentication', () {
+    bool isLogged = false;
     const MethodChannel channel = MethodChannel(
       'app.meedu/flutter_facebook_auth',
     );
@@ -33,10 +32,11 @@ void main() {
             return await MockData.getUserData(fields);
         }
       });
-      facebookAuth = FacebookAuth.instance;
+      facebookAuth = FacebookAuth.i;
     });
 
     test('login request', () async {
+      expect(facebookAuth.isWebSdkInitialized, true);
       expect(await facebookAuth.accessToken, null);
       facebookAuth.webInitialize(appId: "1233443", cookie: true, xfbml: true, version: "v9.0");
       final result = await facebookAuth.login();