+++ facebook_auth_platform_interface/lib/src/method_cahnnel.dart
@@ -109,4 +109,7 @@ class MethodCahnnelFacebookAuth extends FacebookAuthPlatform {
     }
     return null;
   }
+
+  @override
+  bool get isWebSdkInitialized => true;
 }